Power Domain GPIO: Ensure init doesn’t fail if pin configuration fails
When a power domain is connected to another power domain,`pd_gpio_init` attempts to configure the gpio pin to `GPIO_DISCONNECTED`.

But if that failed, the init function returned -EIO, which caused the device to be marked as not initialized / ready.

This meant that it couldn’t be retrieved through `device_get_binding()`.